No longer test on Travis.
We are tested on jenkins.plone.org, and the Travis setup on master is pretty broken.

The master branch has been failing for a long time with this error:

An error occurred while generating the build script.

https://lint.travis-ci.org was complaining about the addons section.
This can be tricky to get right, because the linter errors are not entirely obvious:
- value for addons section is empty, dropping
- in addons section: dropping sauce_connect section: unexpected sequence

In our case, we had dashes in front of the options, and that broke it.

But after fixing that, it was still very broken, missing distributions.